Mike | move to a utility module so that we can use in other places
