+++ src/utility/index.jsx
@@ -6,6 +6,11 @@ import { API_HOST } from 'config';
 import styles from 'components/palette.css';
 import { TableLink, TableHeroImage } from 'components/Table';
 import { KDA } from 'components/Visualizations';
+
+// ------------------------------ RE-EXPORTS ------------------------------
+export { default as bucketizeColumns } from './bucketizeColumns';
+
+// ------------------------------ END RE-EXPORTS ------------------------------
 // TODO - add in the relevant text invocations of TableHeroImage
 export const isRadiant = (playerSlot) => playerSlot < 128;
